+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 - Fixed a bug where documentation would not publish to Hexdocs for packages with
   a number in the name.
 - Fixed a bug where aliased unqualified types and values of the same name could
-  produce an incorrect error. 
+  produce an incorrect error.
 
 ## v0.32.1 - 2023-11-02